feat: group activities by month + style toggle + create Activity component
