+++ src/assets/styles/app.scss
@@ -1,9 +1,13 @@
+@import url("https://fonts.googleapis.com/css2?family=DM+Mono:wght@300&display=swap");
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
 body {
   margin: 0;
+  font-family: "DM Mono", monospace;
+  font-weight: 300;
 }
 
 #app {
