+++ src/components/Activities.svelte
@@ -1,39 +1,36 @@
 <script>
   import { getActivities } from "@/services/api";
   import { onMount } from "svelte";
+  import Activity from "./Activity.svelte";
 
-  let activities = [];
-  let showPrivate = true;
+  let activities = {};
+  let showPrivate = false;
   let page = 1;
   let loading = true;
 
   onMount(() => loadActivities());
 
   const loadActivities = async () => {
-    console.log("load activities");
     const { data } = await getActivities(page);
+
     if (data) {
-      activities = [
+      let newActivities = Object.groupBy(
+        // only display cycling activities
+        data.filter((activity) => activity.type == "Ride"),
+        (activity) => {
+          // groupe activities by month
+          const date = new Date(activity.start_date);
+          return `${date.getMonth() + 1} ${date.getFullYear()}`;
+        }
+      );
+
+      activities = {
         ...activities,
-        ...data.filter((activity) => activity.type == "Ride"),
-      ];
+        ...newActivities,
+      };
+
       loading = false;
     }
-    console.log(activities);
-  };
-
-  const elevationGain = (activity) => {
-    return Math.floor(
-      (activity.total_elevation_gain / activity.distance) * 4000
-    );
-  };
-
-  const speed = (activity) => {
-    return Math.floor(activity.average_speed * 25 - 90);
-  };
-
-  const distance = (activity) => {
-    return Math.floor(activity.distance / 500);
   };
 
   const handleScroll = (e) => {
@@ -49,52 +46,25 @@
 
 <svelte:window on:scroll={handleScroll} />
 
-{#if activities.length > 1}
-  <label
-    class="bg-purple-300 text-purple-900 px-4 py-2 rounded-full fixed left-4 bottom-4"
-  >
-    <input type="checkbox" bind:checked={showPrivate} />
-    Afficher les activités privées ?
-  </label>
+<label
+  class="bg-purple-300 text-purple-900 p-2 pr-4 rounded-full fixed left-4 bottom-4 z-10 flex gap-4"
+>
+  <input type="checkbox" bind:checked={showPrivate} class="sr-only peer" />
+  <div
+    class="relative w-11 h-6 bg-white peer-focus:outline-none peer-focus:ring-0 rounded-full peer peer-checked:after:translate-x-full rtl:peer-checked:after:-translate-x-full peer-checked:after:bg-white peer-checked:bg-purple-800 after:content-[''] after:absolute after:top-[2px] after:start-[2px] after:bg-purple-800 after:border after:rounded-full after:h-5 after:w-5 after:transition-all"
+  ></div>
 
-  <section class="flex flex-col items-center gap-4">
-    {#each activities as activity}
-      <div class="relative {activity.private && !showPrivate ? 'hidden' : ''}">
-        <div
-          class="activity"
-          style="
-            width: {distance(activity)}px;
-            height: {distance(activity)}px;
-            opacity: {speed(activity)}%;"
-        >
-          <div
-            class="activity__elevation-gain"
-            style="height: {elevationGain(activity)}%"
-          ></div>
-        </div>
-        <span
-          class="absolute bg-white top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 w-52 text-center hidden"
-          >{activity.name}</span
-        >
-      </div>
+  <span>Afficher les activités privées ?</span>
+</label>
+
+<section class="flex flex-col items-center gap-4">
+  {#each Object.keys(activities) as activityKey}
+    <h2 class="text-xl my-4 text-purple-500">
+      {activityKey}
+    </h2>
+
+    {#each activities[activityKey] as activity}
+      <Activity {activity} {showPrivate} />
     {/each}
-  </section>
-{/if}
-
-<style lang="scss">
-  .activity {
-    border-radius: 100%;
-    border: 1px solid theme("colors.black");
-    position: relative;
-    overflow: hidden;
-
-    &__elevation-gain {
-      position: absolute;
-      left: 0;
-      bottom: 0;
-      background-color: black;
-      width: 100%;
-      height: 100%;
-    }
-  }
-</style>
+  {/each}
+</section>
