+++ src/components/Login.svelte
@@ -33,6 +33,6 @@
     href={authorizationUrl}
     class="bg-purple-800 text-white px-6 py-4 rounded-full hover:bg-purple-700"
   >
-    Autoriser la connexion avec Strava 🚀
+    Autoriser la connexion avec Strava 🪩
   </a>
 </div>
