+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Login } from "@/components/Login.svelte";
 export { default as Activities } from "@/components/Activities.svelte";
+export { default as Activity } from "@/components/Activity.svelte";
